Issue-153 : Remove the elasticsearch-test in current form - Version 5.0 and above stopped support  for the elasticsearch in-jvm instances. To get around that and test the changes, we will have to run the test against the local version.
