+++ elasticsearch-core/src/test/scala/com/sumologic/elasticsearch/restlastic/RestlasticSearchClientTest.scala
@@ -20,7 +20,6 @@ package com.sumologic.elasticsearch.restlastic
 
 import com.sumologic.elasticsearch.restlastic.RestlasticSearchClient.ReturnTypes._
 import com.sumologic.elasticsearch.restlastic.dsl.Dsl._
-import com.sumologic.elasticsearch_test.ElasticsearchIntegrationTest
 import org.scalatest._
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.time.{Millis, Seconds, Span}
@@ -31,9 +30,8 @@ import org.json4s.native.JsonMethods._
 import scala.concurrent.duration._
 import scala.concurrent.{Await, Future}
 
-class RestlasticSearchClientTest extends WordSpec with Matchers with ScalaFutures with BeforeAndAfterAll
+class RestlasticSearchClientTest extends WordSpec with Matchers with BeforeAndAfterAll
   with ElasticsearchIntegrationTest with OneInstancePerTest {
-  val index = Index(IndexName)
   val tpe = Type("foo")
   val analyzerName = Name("keyword_lowercase")
 
@@ -41,20 +39,11 @@ class RestlasticSearchClientTest extends WordSpec with Matchers with ScalaFuture
 
   implicit val patience = PatienceConfig(timeout = scaled(Span(10, Seconds)), interval = scaled(Span(50, Millis)))
 
-  lazy val restClient = {
-    val (host, port) = endpoint
-    val client = new RestlasticSearchClient(new StaticEndpoint(Endpoint(host, port)))
-    val analyzer = Analyzer(analyzerName, Keyword, Lowercase)
-    val indexSetting = IndexSetting(12, 1, analyzer, 30)
-    val indexFut = client.createIndex(index, Some(indexSetting))
-    indexFut.futureValue
-    client
-  }
+
 
   private def refreshWithClient(): Unit = {
     Await.result(restClient.refresh(index), 2.seconds)
   }
-
 
   "RestlasticSearchClient" should {
     "Be able to create an index and setup index setting with keyword & edgengram lowercase analyzer" in {
