--- elasticsearch-test/.gitignore
@@ -1,12 +1,0 @@
-*.iws
-*.iml
-*.ipr
-*.bak
-*.hprof
-*.diff
-/target/
-.pull.exitcode
-.pull.out
-/data
-/logs
-/temp
