--- elasticsearch-test/pom.xml
@@ -1,64 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>elasticsearch-test</artifactId>
-  <name>elasticsearch-test</name>
-  <packaging>jar</packaging>
-  <version>2.0.7-SNAPSHOT</version>
-  <scm><connection>scm:git:git@github.com:Sanyaku/sumologic.git/tags/all-2.0.7-SNAPSHOT/sumologic.git</connection></scm>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <scala.fatalWarnings>-Xfatal-warnings</scala.fatalWarnings>
-  </properties>
-
-
-  <parent>
-    <artifactId>all</artifactId>
-    <groupId>com.sumologic.elasticsearch</groupId>
-    <version>2.0.7-SNAPSHOT</version>
-    <relativePath>../</relativePath>
-  </parent>
-
-  <dependencies>
-
-    <!-- ATTN: If you need elasticsearch-test you need to also add these dependencies to your pom directly
-         with <scope>test</scope>! -->
-    <dependency>
-      <groupId>org.apache.lucene</groupId>
-      <artifactId>lucene-core</artifactId>
-      <version>${lucene.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.lucene</groupId>
-      <artifactId>lucene-analyzers-common</artifactId>
-      <version>${lucene.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.elasticsearch</groupId>
-      <artifactId>elasticsearch</artifactId>
-      <version>${elasticsearch.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>org.elasticsearch</groupId>
-      <artifactId>elasticsearch</artifactId>
-      <version>${elasticsearch.version}</version>
-      <type>test-jar</type>
-    </dependency>
-
-    <!-- Non-Sumo Test Dependencies. -->
-    <dependency>
-      <groupId>org.scalatest</groupId>
-      <artifactId>scalatest_${scala.version.major}</artifactId>
-      <version>2.2.6</version>
-    </dependency>
-
-  </dependencies>
-
-  <build>
-    <testSourceDirectory>src/test/scala</testSourceDirectory>
-  </build>
-</project>
