+++ pom.xml
@@ -27,8 +27,8 @@
     <properties>
         <scala.version.major>2.11</scala.version.major>
         <akka.version>2.3.11</akka.version>
-        <elasticsearch.version>2.3.5</elasticsearch.version>
-        <lucene.version>5.5.0</lucene.version>
+        <elasticsearch.version>6.1.0</elasticsearch.version>
+        <lucene.version>7.1.0</lucene.version>
     </properties>
     <dependencies>
         <dependency>
